refactor(ver2): split prob-spatial scorers for code health
